GA-102 code cleanup
