+++ frontend/src/api/userApi.js
@@ -1,8 +1,11 @@
+// Base URL — adjust port if needed
+const API_BASE = "http://localhost:3001/api";
+
 export async function registerUserWithJwt(user) {
     try {
-        const res = await fetch('http://localhost:3001/api/users', {
+        const res = await fetch(`${API_BASE}/users`, {
             method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
+            headers: {'Content-Type': 'application/json'},
             body: JSON.stringify(user)
         });
 
@@ -24,23 +27,33 @@ export async function registerUserWithJwt(user) {
     }
 }
 
+/**
+ * POST api/tokens
+ * @param {string} mail of the user, in order to login
+ * @param {string} password of the user, in order to login
+ * @returns {Promise<*|string|number>}
+ * - 400 if the email or password is wrong
+ * - 500 if an unexpected error occurred (catch invoked)
+ * - token if successfully logged in
+ */
 export async function loginWithJwt(mail, password) {
     try {
-        const res = await fetch('http://localhost:3001/api/tokens', {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify({ mail, password })
+        const result = await fetch(`${API_BASE}/tokens`, {
+            method: "POST",
+            body: JSON.stringify({mail, password}),
+            headers: {"Content-Type": "application/json"},
         });
 
-        const data = await res.json();
-
-        if (res.status === 200 && data.token) {
-            return { token: data.token };
-        } else {
-            return { error: data.error || 'Login failed' };
+        if (!result.ok) {
+            console.error("Login error:", result.statusText);
+            return 400;
         }
-    } catch (err) {
-        console.error('Login error:', err);
-        return { error: 'Server error during login' };
+        const data = await result.json();
+        const token = data.token;
+        localStorage.setItem("token", token);
+        return token;
+    } catch (error) {
+        console.error("Network error:", error.message);
+        return 500;
     }
 }
